fix show job proc error

+++ fe/src/main/java/org/apache/doris/load/ExportMgr.java
@@ -277,7 +277,7 @@ public class ExportMgr {
         }
     }
 
-    public Integer getJobNum(ExportJob.JobState state, long dbId) {
+    public long getJobNum(ExportJob.JobState state, long dbId) {
         int size = 0;
         readLock();
         try {
